WIP: initial integration test - multiple include with deep merge
