+++ config/config_partial.go
@@ -1,13 +1,11 @@
 package config
 
 import (
-	"encoding/json"
 	"fmt"
 	"path/filepath"
 
 	"github.com/hashicorp/hcl/v2"
 	"github.com/hashicorp/hcl/v2/hclparse"
-	"github.com/hashicorp/hcl/v2/hclwrite"
 	"github.com/zclconf/go-cty/cty"
 
 	"github.com/gruntwork-io/terragrunt/errors"
@@ -333,177 +331,17 @@ func partialParseIncludedConfig(includedConfig *IncludeConfig, terragruntOptions
 // the config.
 // For consistency, `include` in the call to `decodeHcl` is always assumed to be nil. Either it really is nil (parsing
 // the child config), or it shouldn't be used anyway (the parent config shouldn't have an include block).
-//
-// We take a two pass approach to parsing include blocks to support include blocks without a label. Ideally we can parse
-// include blocks with and without labels in a single pass, but the HCL parser is fairly restrictive when it comes to
-// parsing blocks with labels, requiring the exact number of expected labels in the parsing step.
-// To handle this restriction, we first see if there are any include blocks without any labels, and if there is, we
-// modify it in the file object to inject the label as "".
 func decodeAsTerragruntInclude(
 	file *hcl.File,
 	filename string,
 	terragruntOptions *options.TerragruntOptions,
 	extensions EvalContextExtensions,
 ) ([]IncludeConfig, error) {
-	updatedBytes, isUpdated, err := updateBareIncludeBlock(file, filename)
-	if err != nil {
-		return nil, err
-	}
-	if isUpdated {
-		// Code was updated, so we need to reparse the new updated contents. This is necessarily because the blocks
-		// returned by hclparse does not support editing, and so we have to go through hclwrite, which leads to a
-		// different AST representation.
-		file, err = parseHcl(hclparse.NewParser(), string(updatedBytes), filename)
-		if err != nil {
-			return nil, err
-		}
-	}
-
 	tgInc := terragruntIncludeMultiple{}
 	if err := decodeHcl(file, filename, &tgInc, terragruntOptions, extensions); err != nil {
 		return nil, err
 	}
 	return tgInc.Include, nil
-}
-
-// updateBareIncludeBlock searches the parsed terragrunt contents for a bare include block (include without a label),
-// and convert it to one with empty string as the label. This is necessary because the hcl parser is strictly enforces
-// label counts when parsing out labels with a go struct.
-//
-// Returns the updated contents, a boolean indicated whether anything changed, and an error (if any).
-func updateBareIncludeBlock(file *hcl.File, filename string) ([]byte, bool, error) {
-	if filepath.Ext(filename) == ".json" {
-		return updateBareIncludeBlockJSON(file.Bytes)
-	}
-
-	hclFile, err := hclwrite.ParseConfig(file.Bytes, filename, hcl.InitialPos)
-	if err != nil {
-		return nil, false, errors.WithStackTrace(err)
-	}
-
-	codeWasUpdated := false
-	for _, block := range hclFile.Body().Blocks() {
-		if block.Type() == "include" && len(block.Labels()) == 0 {
-			if codeWasUpdated {
-				return nil, false, errors.WithStackTrace(MultipleBareIncludeBlocksErr{})
-			}
-			block.SetLabels([]string{""})
-			codeWasUpdated = true
-		}
-	}
-	return hclFile.Bytes(), codeWasUpdated, nil
-}
-
-// updateBareIncludeBlockJSON implements the logic for updateBareIncludeBlock when the terragrunt.hcl configuration is
-// encoded in json. The json version of this function is fairly complex due to the flexibility in how the blocks are
-// encoded. That is, all of the following are valid encodings of a terragrunt.hcl.json file that has a bare include
-// block:
-//
-// Case 1: a single include block as top level:
-// {
-//   "include": {
-//     "path": "foo"
-//   }
-// }
-//
-// Case 2: a single include block in list:
-// {
-//   "include": [
-//     {"path": "foo"}
-//   ]
-// }
-//
-// Case 3: mixed bare and labeled include block as list:
-// {
-//   "include": [
-//     {"path": "foo"},
-//     {
-//       "labeled": {"path": "bar"}
-//     }
-//   ]
-// }
-//
-// For simplicity of implementation, we focus on handling Case 1 and 2, and ignore Case 3. If we see Case 3, we will
-// error out. Instead, the user should handle this case explicitly using the object encoding instead of list encoding:
-// {
-//   "include": {
-//     "": {"path": "foo"},
-//     "labeled": {"path": "bar"}
-//   }
-// }
-// If the multiple include blocks are encoded in this way in the json configuration, nothing needs to be done by this
-// function.
-func updateBareIncludeBlockJSON(fileBytes []byte) ([]byte, bool, error) {
-	var parsed map[string]interface{}
-	if err := json.Unmarshal(fileBytes, &parsed); err != nil {
-		return nil, false, errors.WithStackTrace(err)
-	}
-	includeBlock, hasKey := parsed["include"]
-	if !hasKey {
-		// No include block, so don't do anything
-		return fileBytes, false, nil
-	}
-	switch typed := includeBlock.(type) {
-	case []interface{}:
-		if len(typed) == 0 {
-			// No include block, so don't do anything
-			return nil, false, nil
-		} else if len(typed) > 1 {
-			// Could be multiple bare includes, or Case 3. We simplify the handling of this case by erroring out,
-			// ignoring the possibility of Case 3, which, while valid HCL encoding, is too complex to detect and handle
-			// here. Instead we will recommend users use the object encoding.
-			return nil, false, errors.WithStackTrace(MultipleBareIncludeBlocksErr{})
-		}
-
-		// Make sure this is Case 2, and not Case 3 with a single labeled block. If Case 2, update to inject the labeled
-		// version. Otherwise, return without modifying.
-		singleBlock := typed[0]
-		if jsonIsIncludeBlock(singleBlock) {
-			return updateSingleBareIncludeInParsedJSON(parsed, singleBlock)
-		}
-		return nil, false, nil
-	case map[string]interface{}:
-		if len(typed) == 0 {
-			// No include block, so don't do anything
-			return nil, false, nil
-		}
-
-		// We will only update the include block if we detect the object to represent an include block. Otherwise, the
-		// blocks are labeled so we can pass forward to the tg parser step.
-		if jsonIsIncludeBlock(typed) {
-			return updateSingleBareIncludeInParsedJSON(parsed, typed)
-		}
-		return nil, false, nil
-	}
-
-	return nil, false, errors.WithStackTrace(IncludeIsNotABlockErr{parsed: includeBlock})
-}
-
-// updateBareIncludeInParsedJSON replaces the include attribute into a block with the label "" in the json. Note that we
-// can directly assign to the map with the single "" key without worrying about the possibility of other include blocks
-// since we will only call this function if there is only one include block, and that is a bare block with no labels.
-func updateSingleBareIncludeInParsedJSON(parsed map[string]interface{}, newVal interface{}) ([]byte, bool, error) {
-	parsed["include"] = map[string]interface{}{"": newVal}
-	updatedBytes, err := json.Marshal(parsed)
-	return updatedBytes, true, errors.WithStackTrace(err)
-}
-
-// jsonIsIncludeBlock checks if the arbitrary json data is the include block. The data is determined to be an include
-// block if:
-// - It is an object
-// - Has the 'path' attribute
-// - The 'path' attribute is a string
-func jsonIsIncludeBlock(jsonData interface{}) bool {
-	typed, isMap := jsonData.(map[string]interface{})
-	if isMap {
-		pathAttr, hasPath := typed["path"]
-		if hasPath {
-			_, pathIsString := pathAttr.(string)
-			return pathIsString
-		}
-		return false
-	}
-	return false
 }
 
 // Custom error types
@@ -514,18 +352,4 @@ type InvalidPartialBlockName struct {
 
 func (err InvalidPartialBlockName) Error() string {
 	return fmt.Sprintf("Unrecognized partial block code %d. This is most likely an error in terragrunt. Please file a bug report on the project repository.", err.sectionCode)
-}
-
-type MultipleBareIncludeBlocksErr struct{}
-
-func (err MultipleBareIncludeBlocksErr) Error() string {
-	return "Multiple bare include blocks (include blocks without label) is not supported."
-}
-
-type IncludeIsNotABlockErr struct {
-	parsed interface{}
-}
-
-func (err IncludeIsNotABlockErr) Error() string {
-	return fmt.Sprintf("Parsed include is not a block: %v", err.parsed)
 }
