+++ config/hcl_parser.go
@@ -40,6 +40,12 @@ func parseHcl(parser *hclparse.Parser, hcl string, filename string) (file *hcl.F
 }
 
 // decodeHcl uses the HCL2 parser to decode the parsed HCL into the struct specified by out.
+//
+// Note that we take a two pass approach to support parsing include blocks without a label. Ideally we can parse include
+// blocks with and without labels in a single pass, but the HCL parser is fairly restrictive when it comes to parsing
+// blocks with labels, requiring the exact number of expected labels in the parsing step.  To handle this restriction,
+// we first see if there are any include blocks without any labels, and if there is, we modify it in the file object to
+// inject the label as "".
 func decodeHcl(
 	file *hcl.File,
 	filename string,
@@ -54,6 +60,21 @@ func decodeHcl(
 			err = errors.WithStackTrace(PanicWhileParsingConfig{RecoveredValue: recovered, ConfigFile: filename})
 		}
 	}()
+
+	// Check if we need to update the file to label any bare include blocks.
+	updatedBytes, isUpdated, err := updateBareIncludeBlock(file, filename)
+	if err != nil {
+		return err
+	}
+	if isUpdated {
+		// Code was updated, so we need to reparse the new updated contents. This is necessarily because the blocks
+		// returned by hclparse does not support editing, and so we have to go through hclwrite, which leads to a
+		// different AST representation.
+		file, err = parseHcl(hclparse.NewParser(), string(updatedBytes), filename)
+		if err != nil {
+			return err
+		}
+	}
 
 	evalContext, err := CreateTerragruntEvalContext(filename, terragruntOptions, extensions)
 	if err != nil {
