+++ test/fixture-include-multiple/deep-merge-overlapping/vpc/terragrunt.hcl
@@ -1,0 +1,3 @@
+terraform {
+  source = "${get_terragrunt_dir()}/../../modules/empty"
+}
