+++ test/fixture-include-multiple/modules/empty/main.tf
@@ -1,0 +1,1 @@
+# Intentionally empty
