+++ test/fixture-include-multiple/terragrunt_inputs_override.hcl
@@ -1,0 +1,8 @@
+inputs = {
+  attribute     = "will be replaced"
+  new_attribute = "new val"
+  list_attr     = ["mock"]
+  map_attr = {
+    bar = "baz"
+  }
+}
