+++ test/fixture-include-multiple/terragrunt_vpc_dep.hcl
@@ -1,0 +1,14 @@
+dependency "vpc" {
+  # This will get overridden by child terragrunt.hcl configs
+  config_path = ""
+
+  mock_outputs = {
+    attribute     = "hello"
+    old_attribute = "old val"
+    list_attr     = ["hello"]
+    map_attr = {
+      foo = "bar"
+    }
+  }
+  mock_outputs_allowed_terraform_commands = ["apply", "plan", "destroy", "output"]
+}
