+++ test/fixture-include-multiple/terragrunt_vpc_dep_override.hcl
@@ -1,0 +1,11 @@
+dependency "vpc" {
+  config_path = ""
+  mock_outputs = {
+    attribute     = "will be replaced"
+    new_attribute = "new val"
+    list_attr     = ["mock"]
+    map_attr = {
+      bar = "baz"
+    }
+  }
+}
